Add an HAProxy based edge router to proxy Marathon tasks
This is similar to servicerouter but rather than operate on service
ports it only exposes services / ports in the edge router which are
explicitly requested through setting environment variables for
specific Marathon port indexes.

There are two proxies which edgerouter will setup

1) Virtual Host style proxying via the EDGEROUTER_{n}_VHOST
environment variables. the HAProxy by default always listens on port
80 and it will listen with HTTPS on port 443 if an SSL certificate is
given. EDGEROUTER_{n}_VHOST will look at the hostname of the request
and then forward it to the appropriate Marathon task port.

2) TCP proxying via the EDGEROUTER_{n}_PORT. Note if PORT is given
along with VHOST then it will be HAProxy HTTP proxying. Adds an
additional port which HAProxy binds to and forwards traffic to the
appropriate Marathon task.